admin almose done

+++ app/admin/_components/tableHoc/actions.tsx
@@ -5,38 +5,46 @@ import { useState } from "react";
 
 export type CellActionProps<T> = {
   data: T;
-  onEditClick: () => void;
-  handleDelete: () => void;
-};
+} & ({ editable: true; onEditClick: () => void } | { editable: false }) &
+  ({ deleteable: true; handleDelete: () => void } | { deleteable: false });
 
 function CellAction<T = Record<string, any> & { id: number }>(props: CellActionProps<T>) {
   const [open, setOpen] = useState(false);
 
-  const onDeleteClick = () => setOpen(true);
+  const onDeleteClick = () => {
+    if (props.deleteable) setOpen(true);
+  };
 
   const onConfirmDelete = () => {
-    props.handleDelete();
+    if (props.deleteable) props.handleDelete();
     setOpen(false);
   };
 
   return (
     <>
-      <AlertModal
-        isOpen={open}
-        onClose={() => setOpen(false)}
-        onConfirm={onConfirmDelete}
-        loading={false}
-        okText="Delete"
-        title="Are you sure you want to delete this ?"
-      />
-      <div className="flex items-center gap-4">
-        <div className="flex cursor-pointer items-center hover:text-primary" onClick={props.onEditClick}>
-          <Edit className="mr-2 h-4 w-4" /> Update
-        </div>
+      {props.deleteable ? (
+        <AlertModal
+          isOpen={open}
+          onClose={() => setOpen(false)}
+          onConfirm={onConfirmDelete}
+          loading={false}
+          okText="Delete"
+          title="Are you sure you want to delete this ?"
+        />
+      ) : null}
 
-        <div className="flex cursor-pointer items-center hover:text-primary" onClick={onDeleteClick}>
-          <Trash className="mr-2 h-4 w-4" /> Delete
-        </div>
+      <div className="flex items-center gap-4">
+        {props.editable ? (
+          <div className="flex cursor-pointer items-center hover:text-primary" onClick={props.onEditClick}>
+            <Edit className="mr-2 h-4 w-4" /> Update
+          </div>
+        ) : null}
+
+        {props.deleteable ? (
+          <div className="flex cursor-pointer items-center hover:text-primary" onClick={onDeleteClick}>
+            <Trash className="mr-2 h-4 w-4" /> Delete
+          </div>
+        ) : null}
       </div>
     </>
   );
